+++ database/migrations/2025_02_26_015512_create_sub_tasks_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('task_id');
             $table->foreign('task_id')->references('id')->on('tasks');
             $table->string('content');
-            $table->integer('percentage')->nullable();
+            $table->decimal('percentage', 5, 2)->nullable();
             $table->timestamps();
         });
     }
